refactor: move WiFi logic into helper
